pyproject.toml
[project]
name = "python-game-jam-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pygame-ce>=2.5.3",
    "pytmx>=3.32",
]
